Verify and Debug Segment Integration
Ensure proper initialization of the Segment library using the correct Write Key. Add debugging logs to confirm the library is loaded and check event firing timing by wrapping events in `analytics.ready()`. Utilize the Segment Debugger to verify if events are being received accurately.
[skip gpt_engineer]
